requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.7.4
pydantic-settings==2.12.0
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
odt2md==0.1.0
docx2md==1.0.4
mammoth==1.6.0
markitdown[pdf]==0.1.3
openai==1.54.5
convertapi==2.0.0
httpx==0.27.0
markdown==3.7
python-docx==1.2.0
beautifulsoup4==4.12.3
docx-parser-converter==0.5.1.2
html_to_markdown==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

